created script to create project and app at once, updated remove script to handle files not existing

+++ bin/remove-sample.sh
@@ -1,7 +1,9 @@
 #! /bin/bash
 
-rm  manage.py
+set -eou pipefail
+
+if [ -f manage.py ]; then rm manage.py; fi
 rm -rf dead_songs
 rm -rf api
 
-rm ./Backups/dead_songs.sql
+if [ -f ./Backups/dead_songs.sql ]; then rm ./Backups/dead_songs.sql; fi
